Refactor : 패키지 2차 리팩토링
