+++ src/main/java/ssu/cromi/teamit/repository/findproject/ProjectApplicationRepository.java
@@ -1,4 +1,4 @@
-package ssu.cromi.teamit.repository;
+package ssu.cromi.teamit.repository.findproject;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import ssu.cromi.teamit.entity.findproject.ProjectApplication;
