+++ src/main/java/ssu/cromi/teamit/repository/teamup/ProjectRepository.java
@@ -1,6 +1,6 @@
 // DB 접근
 
-package ssu.cromi.teamit.repository;
+package ssu.cromi.teamit.repository.teamup;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import ssu.cromi.teamit.entity.teamup.Project;
